getting first cell id address, subcity and maincity
