--- filtered/filtered_sample_fixed.csv
@@ -1,4 +1,0 @@
-called_to,call_duration,call_time
-9123456780,45,2025-04-15 14:20:00
-9234567890,120,2025-04-15 15:30:00
-9321654987,200,2025-04-15 19:40:00
